+++ tests/Mfa/Physical/SmsTest.php
@@ -91,16 +91,20 @@ class SmsTest extends TestCase
 
 
     }
+    /** @test */
+    public function returns_true()
+    {
+        $this->assertTrue(true);
+    }
+    
 
     /**
-     * @test
      * @group physical
      */
     public function adding_sms_factor_with_challenge_flag_send_a_text_message()
     {
 
         $smsFactor = SmsFactor::instantiate();
-//        $smsFactor->phone = '5133608855';
         $smsFactor->challenge = 'Validate with ${code}';
 
         $factor = self::$account->addFactor($smsFactor);
@@ -127,7 +131,7 @@ class SmsTest extends TestCase
                 'middleName' => 'Middle Name',
                 'surname' => 'Surname',
                 'username' => makeUniqueName('AccountTest smsFactor') . 'username',
-                'email' => makeUniqueName('AccountTest smsFactor') .'@mailinator.com',
+                'email' => makeUniqueName('AccountTest smsFactor') .'@testmail.stormpath.com',
                 'password' => 'superP4ss!'
             ])
         );